requirements.txt
streamlit==1.42.0
Authlib

google-cloud-firestore
google-auth

langchain
langchain-community
langchain-core
langchain-openai
langchain-groq

pydantic 
typing

# For Slackbot
python-dotenv
slack_sdk

google-api-python-client
google-cloud
google-auth

langgraph